Console.log personality objects
Theory about why matching logic is producing unexpected outcomes:
The user's personality profile object has its traits in a different
order than that of the candidate object's. The candidate's traits are
in reverse order from that of userPersonality. Furthermore, the addition
of the name, twitter, and imageUrl keys may be a problem